move auth header to nav folder
